requirements.txt
opencv-python~=4.5.5.64
numpy~=1.22.3
Pillow~=9.0.1
pip~=22.0.4
certifi~=2021.10.8
six~=1.16.0
python-dateutil~=2.8.2
matplotlib~=3.5.1
idna~=3.3
urllib3~=1.26.9
fonttools~=4.31.2
pyparsing~=3.0.7
cycler~=0.11.0
packaging~=21.3
requests~=2.27.1
click~=8.0.4
colorama~=0.4.4